--- a/pkg/controller/replication/replication_controller_test.go
+++ b/pkg/controller/replication/replication_controller_test.go
@@ -137,8 +137,8 @@ func TestSyncReplicationControllerDoesNothing(t *testing.T) {
 
 	// 2 running pods, a controller with 2 replicas, sync is a no-op
 	controllerSpec := newReplicationController(2)
-	manager.rcStore.Store.Add(controllerSpec)
-	newPodList(manager.podStore.Store, 2, api.PodRunning, controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
+	newPodList(manager.podStore.Indexer, 2, api.PodRunning, controllerSpec)
 
 	manager.podControl = &fakePodControl
 	manager.syncReplicationController(getKey(controllerSpec, t))
@@ -154,8 +154,8 @@ func TestSyncReplicationControllerDeletes(t *testing.T) {
 
 	// 2 running pods and a controller with 1 replica, one pod delete expected
 	controllerSpec := newReplicationController(1)
-	manager.rcStore.Store.Add(controllerSpec)
-	newPodList(manager.podStore.Store, 2, api.PodRunning, controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
+	newPodList(manager.podStore.Indexer, 2, api.PodRunning, controllerSpec)
 
 	manager.syncReplicationController(getKey(controllerSpec, t))
 	validateSyncReplication(t, &fakePodControl, 0, 1)
@@ -177,7 +177,7 @@ func TestDeleteFinalStateUnknown(t *testing.T) {
 	// The DeletedFinalStateUnknown object should cause the rc manager to insert
 	// the controller matching the selectors of the deleted pod into the work queue.
 	controllerSpec := newReplicationController(1)
-	manager.rcStore.Store.Add(controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
 	pods := newPodList(nil, 1, api.PodRunning, controllerSpec)
 	manager.deletePod(cache.DeletedFinalStateUnknown{Key: "foo", Obj: &pods.Items[0]})
 
@@ -201,7 +201,7 @@ func TestSyncReplicationControllerCreates(t *testing.T) {
 
 	// A controller with 2 replicas and no pods in the store, 2 creates expected
 	rc := newReplicationController(2)
-	manager.rcStore.Store.Add(rc)
+	manager.rcStore.Indexer.Add(rc)
 
 	fakePodControl := controller.FakePodControl{}
 	manager.podControl = &fakePodControl
@@ -225,9 +225,9 @@ func TestStatusUpdatesWithoutReplicasChange(t *testing.T) {
 	// Steady state for the replication controller, no Status.Replicas updates expected
 	activePods := 5
 	rc := newReplicationController(activePods)
-	manager.rcStore.Store.Add(rc)
+	manager.rcStore.Indexer.Add(rc)
 	rc.Status = api.ReplicationControllerStatus{Replicas: activePods}
-	newPodList(manager.podStore.Store, activePods, api.PodRunning, rc)
+	newPodList(manager.podStore.Indexer, activePods, api.PodRunning, rc)
 
 	fakePodControl := controller.FakePodControl{}
 	manager.podControl = &fakePodControl
@@ -268,10 +268,10 @@ func TestControllerUpdateReplicas(t *testing.T) {
 	// Insufficient number of pods in the system, and Status.Replicas is wrong;
 	// Status.Replica should update to match number of pods in system, 1 new pod should be created.
 	rc := newReplicationController(5)
-	manager.rcStore.Store.Add(rc)
+	manager.rcStore.Indexer.Add(rc)
 	rc.Status = api.ReplicationControllerStatus{Replicas: 2, ObservedGeneration: 0}
 	rc.Generation = 1
-	newPodList(manager.podStore.Store, 4, api.PodRunning, rc)
+	newPodList(manager.podStore.Indexer, 4, api.PodRunning, rc)
 
 	// This response body is just so we don't err out decoding the http response
 	response := runtime.EncodeOrDie(testapi.Default.Codec(), &api.ReplicationController{})
@@ -308,8 +308,8 @@ func TestSyncReplicationControllerDormancy(t *testing.T) {
 	manager.podControl = &fakePodControl
 
 	controllerSpec := newReplicationController(2)
-	manager.rcStore.Store.Add(controllerSpec)
-	newPodList(manager.podStore.Store, 1, api.PodRunning, controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
+	newPodList(manager.podStore.Indexer, 1, api.PodRunning, controllerSpec)
 
 	// Creates a replica and sets expectations
 	controllerSpec.Status.Replicas = 1
@@ -428,7 +428,7 @@ func TestWatchControllers(t *testing.T) {
 	// and closes the received channel to indicate that the test can finish.
 	manager.syncHandler = func(key string) error {
 
-		obj, exists, err := manager.rcStore.Store.GetByKey(key)
+		obj, exists, err := manager.rcStore.Indexer.GetByKey(key)
 		if !exists || err != nil {
 			t.Errorf("Expected to find controller under key %v", key)
 		}
@@ -465,13 +465,13 @@ func TestWatchPods(t *testing.T) {
 
 	// Put one rc and one pod into the controller's stores
 	testControllerSpec := newReplicationController(1)
-	manager.rcStore.Store.Add(testControllerSpec)
+	manager.rcStore.Indexer.Add(testControllerSpec)
 	received := make(chan string)
 	// The pod update sent through the fakeWatcher should figure out the managing rc and
 	// send it into the syncHandler.
 	manager.syncHandler = func(key string) error {
 
-		obj, exists, err := manager.rcStore.Store.GetByKey(key)
+		obj, exists, err := manager.rcStore.Indexer.GetByKey(key)
 		if !exists || err != nil {
 			t.Errorf("Expected to find controller under key %v", key)
 		}
@@ -508,7 +508,7 @@ func TestUpdatePods(t *testing.T) {
 	received := make(chan string)
 
 	manager.syncHandler = func(key string) error {
-		obj, exists, err := manager.rcStore.Store.GetByKey(key)
+		obj, exists, err := manager.rcStore.Indexer.GetByKey(key)
 		if !exists || err != nil {
 			t.Errorf("Expected to find controller under key %v", key)
 		}
@@ -522,14 +522,14 @@ func TestUpdatePods(t *testing.T) {
 
 	// Put 2 rcs and one pod into the controller's stores
 	testControllerSpec1 := newReplicationController(1)
-	manager.rcStore.Store.Add(testControllerSpec1)
+	manager.rcStore.Indexer.Add(testControllerSpec1)
 	testControllerSpec2 := *testControllerSpec1
 	testControllerSpec2.Spec.Selector = map[string]string{"bar": "foo"}
 	testControllerSpec2.Name = "barfoo"
-	manager.rcStore.Store.Add(&testControllerSpec2)
+	manager.rcStore.Indexer.Add(&testControllerSpec2)
 
 	// Put one pod in the podStore
-	pod1 := newPodList(manager.podStore.Store, 1, api.PodRunning, testControllerSpec1).Items[0]
+	pod1 := newPodList(manager.podStore.Indexer, 1, api.PodRunning, testControllerSpec1).Items[0]
 	pod2 := pod1
 	pod2.Labels = testControllerSpec2.Spec.Selector
 
@@ -566,9 +566,9 @@ func TestControllerUpdateRequeue(t *testing.T) {
 	manager.podStoreSynced = alwaysReady
 
 	rc := newReplicationController(1)
-	manager.rcStore.Store.Add(rc)
+	manager.rcStore.Indexer.Add(rc)
 	rc.Status = api.ReplicationControllerStatus{Replicas: 2}
-	newPodList(manager.podStore.Store, 1, api.PodRunning, rc)
+	newPodList(manager.podStore.Indexer, 1, api.PodRunning, rc)
 
 	fakePodControl := controller.FakePodControl{}
 	manager.podControl = &fakePodControl
@@ -648,7 +648,7 @@ func doTestControllerBurstReplicas(t *testing.T, burstReplicas, numReplicas int)
 	manager.podControl = &fakePodControl
 
 	controllerSpec := newReplicationController(numReplicas)
-	manager.rcStore.Store.Add(controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
 
 	expectedPods := 0
 	pods := newPodList(nil, numReplicas, api.PodPending, controllerSpec)
@@ -662,14 +662,14 @@ func doTestControllerBurstReplicas(t *testing.T, burstReplicas, numReplicas int)
 	for _, replicas := range []int{numReplicas, 0} {
 
 		controllerSpec.Spec.Replicas = replicas
-		manager.rcStore.Store.Add(controllerSpec)
+		manager.rcStore.Indexer.Add(controllerSpec)
 
 		for i := 0; i < numReplicas; i += burstReplicas {
 			manager.syncReplicationController(getKey(controllerSpec, t))
 
 			// The store accrues active pods. It's also used by the rc to determine how many
 			// replicas to create.
-			activePods := len(manager.podStore.Store.List())
+			activePods := len(manager.podStore.Indexer.List())
 			if replicas != 0 {
 				// This is the number of pods currently "in flight". They were created by the rc manager above,
 				// which then puts the rc to sleep till all of them have been observed.
@@ -683,7 +683,7 @@ func doTestControllerBurstReplicas(t *testing.T, burstReplicas, numReplicas int)
 				// This simulates the watch events for all but 1 of the expected pods.
 				// None of these should wake the controller because it has expectations==BurstReplicas.
 				for i := 0; i < expectedPods-1; i++ {
-					manager.podStore.Store.Add(&pods.Items[i])
+					manager.podStore.Indexer.Add(&pods.Items[i])
 					manager.addPod(&pods.Items[i])
 				}
 
@@ -701,7 +701,7 @@ func doTestControllerBurstReplicas(t *testing.T, burstReplicas, numReplicas int)
 				}
 				validateSyncReplication(t, &fakePodControl, 0, expectedPods)
 				for i := 0; i < expectedPods-1; i++ {
-					manager.podStore.Store.Delete(&pods.Items[i])
+					manager.podStore.Indexer.Delete(&pods.Items[i])
 					manager.deletePod(&pods.Items[i])
 				}
 				podExp, exists, err := manager.expectations.GetExpectations(rcKey)
@@ -722,17 +722,17 @@ func doTestControllerBurstReplicas(t *testing.T, burstReplicas, numReplicas int)
 			// The last add pod will decrease the expectation of the rc to 0,
 			// which will cause it to create/delete the remaining replicas up to burstReplicas.
 			if replicas != 0 {
-				manager.podStore.Store.Add(&pods.Items[expectedPods-1])
+				manager.podStore.Indexer.Add(&pods.Items[expectedPods-1])
 				manager.addPod(&pods.Items[expectedPods-1])
 			} else {
-				manager.podStore.Store.Delete(&pods.Items[expectedPods-1])
+				manager.podStore.Indexer.Delete(&pods.Items[expectedPods-1])
 				manager.deletePod(&pods.Items[expectedPods-1])
 			}
 			pods.Items = pods.Items[expectedPods:]
 		}
 
 		// Confirm that we've created the right number of replicas
-		activePods := len(manager.podStore.Store.List())
+		activePods := len(manager.podStore.Indexer.List())
 		if activePods != controllerSpec.Spec.Replicas {
 			t.Fatalf("Unexpected number of active pods, expected %d, got %d", controllerSpec.Spec.Replicas, activePods)
 		}
@@ -768,9 +768,9 @@ func TestRCSyncExpectations(t *testing.T) {
 	manager.podControl = &fakePodControl
 
 	controllerSpec := newReplicationController(2)
-	manager.rcStore.Store.Add(controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
 	pods := newPodList(nil, 2, api.PodPending, controllerSpec)
-	manager.podStore.Store.Add(&pods.Items[0])
+	manager.podStore.Indexer.Add(&pods.Items[0])
 	postExpectationsPod := pods.Items[1]
 
 	manager.expectations = FakeRCExpectations{
@@ -778,7 +778,7 @@ func TestRCSyncExpectations(t *testing.T) {
 			// If we check active pods before checking expectataions, the rc
 			// will create a new replica because it doesn't see this pod, but
 			// has fulfilled its expectations.
-			manager.podStore.Store.Add(&postExpectationsPod)
+			manager.podStore.Indexer.Add(&postExpectationsPod)
 		},
 	}
 	manager.syncReplicationController(getKey(controllerSpec, t))
@@ -791,7 +791,7 @@ func TestDeleteControllerAndExpectations(t *testing.T) {
 	manager.podStoreSynced = alwaysReady
 
 	rc := newReplicationController(1)
-	manager.rcStore.Store.Add(rc)
+	manager.rcStore.Indexer.Add(rc)
 
 	fakePodControl := controller.FakePodControl{}
 	manager.podControl = &fakePodControl
@@ -822,7 +822,7 @@ func TestDeleteControllerAndExpectations(t *testing.T) {
 
 	// This should have no effect, since we've deleted the rc.
 	podExp.Seen(1, 0)
-	manager.podStore.Store.Replace(make([]interface{}, 0), "0")
+	manager.podStore.Indexer.Replace(make([]interface{}, 0), "0")
 	manager.syncReplicationController(getKey(rc, t))
 	validateSyncReplication(t, &fakePodControl, 0, 0)
 }
@@ -838,7 +838,7 @@ func TestRCManagerNotReady(t *testing.T) {
 	// want to end up creating replicas in this case until the pod reflector
 	// has synced, so the rc manager should just requeue the rc.
 	controllerSpec := newReplicationController(1)
-	manager.rcStore.Store.Add(controllerSpec)
+	manager.rcStore.Indexer.Add(controllerSpec)
 
 	rcKey := getKey(controllerSpec, t)
 	manager.syncReplicationController(rcKey)
@@ -881,7 +881,7 @@ func TestOverlappingRCs(t *testing.T) {
 		}
 		shuffledControllers := shuffle(controllers)
 		for j := range shuffledControllers {
-			manager.rcStore.Store.Add(shuffledControllers[j])
+			manager.rcStore.Indexer.Add(shuffledControllers[j])
 		}
 		// Add a pod and make sure only the oldest rc is synced
 		pods := newPodList(nil, 1, api.PodPending, controllers[0])
@@ -910,11 +910,60 @@ func TestRCManagerInit(t *testing.T) {
 
 	client := client.NewOrDie(&client.Config{Host: testServer.URL, ContentConfig: client.ContentConfig{GroupVersion: testapi.Default.GroupVersion()}})
 	manager := NewReplicationManager(client, controller.NoResyncPeriodFunc, BurstReplicas)
-	manager.rcStore.Store.Add(rc)
+	manager.rcStore.Indexer.Add(rc)
 	manager.podStoreSynced = alwaysReady
-	controller.SyncAllPodsWithStore(manager.kubeClient, manager.podStore.Store)
+	controller.SyncAllPodsWithStore(manager.kubeClient, manager.podStore.Indexer)
 	fakePodControl := &controller.FakePodControl{}
 	manager.podControl = fakePodControl
 	manager.syncReplicationController(getKey(rc, t))
 	validateSyncReplication(t, fakePodControl, 0, 0)
+}
+
+func BenchmarkGetPodControllers(b *testing.B) {
+	client := client.NewOrDie(&client.Config{Host: "", ContentConfig: client.ContentConfig{GroupVersion: testapi.Default.GroupVersion()}})
+	manager := NewReplicationManager(client, controller.NoResyncPeriodFunc, BurstReplicas)
+
+	const nsNum = 1000
+
+	pods := []api.Pod{}
+	for i := 0; i < nsNum; i++ {
+		ns := fmt.Sprintf("ns-%d", i)
+		for j := 0; j < 10; j++ {
+			rcName := fmt.Sprintf("rc-%d", j)
+			for k := 0; k < 10; k++ {
+				podName := fmt.Sprintf("pod-%d-%d", j, k)
+				pods = append(pods, api.Pod{
+					ObjectMeta: api.ObjectMeta{
+						Name:      podName,
+						Namespace: ns,
+						Labels:    map[string]string{"rcName": rcName},
+					},
+				})
+			}
+		}
+	}
+
+	for i := 0; i < nsNum; i++ {
+		ns := fmt.Sprintf("ns-%d", i)
+		for j := 0; j < 10; j++ {
+			rcName := fmt.Sprintf("rc-%d", j)
+			manager.rcStore.Add(&api.ReplicationController{
+				ObjectMeta: api.ObjectMeta{Name: rcName, Namespace: ns},
+				Spec: api.ReplicationControllerSpec{
+					Selector: map[string]string{"rcName": rcName},
+				},
+			})
+		}
+	}
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		for _, pod := range pods {
+			_, err := manager.rcStore.GetPodControllers(&pod)
+			if err != nil {
+				b.Errorf("unexpected error: %v", err)
+			}
+		}
+	}
 }