add namespace index for cache in replication controller manager