@davidopp 2015-11-25
Also @wojtek-t might be interested.