@lavalamp 2015-11-30
Do you have some before/after benchmark numbers or CPU usage numbers? I want to make sure this is worth the added complexity.

Also there is no testing; what will keep this index in and working in the future?