@mqliang 2015-12-01
@lavalamp I will test the CPU usage soon and rework this PR according to your suggestion if the test indicate the namespace index could improve efficiency.