@mqliang 2015-12-05
@lavalamp 
Result without my PR:
```
BenchmarkGetPodControllers	       1	1398850999 ns/op
ok  	k8s.io/kubernetes/pkg/controller/replication	11.174s
```
Result with my PR:
```
BenchmarkGetPodControllers	       1	2083932426 ns/op
ok  	k8s.io/kubernetes/pkg/controller/replication	11.853s
```

Benchmark code is as follows: 
```
func BenchmarkGetPodControllers(b *testing.B) {
	client := client.NewOrDie(&client.Config{Host: "", GroupVersion: testapi.Default.GroupVersion()})
	manager := NewReplicationManager(client, controller.NoResyncPeriodFunc, BurstReplicas)

	b.ResetTimer()

	for i := 0; i < 100; i++ {
		ns := fmt.Sprintf("ns-%d", i)
		for j:= 0; j < 10; j++ {
			rcName := fmt.Sprintf("rc-%d", j)
			manager.rcStore.Add(&api.ReplicationController{
				ObjectMeta: api.ObjectMeta{Name: rcName, Namespace: ns},
				Spec: api.ReplicationControllerSpec{
					Selector: map[string]string{"rcName": rcName},
				},
			})
		}
	}

	pods := []api.Pod{}
	for i := 0; i < 100; i++ {
		ns := fmt.Sprintf("ns-%d", i)
		for j := 0; j < 10; j++ {
			rcName := fmt.Sprintf("rc-%d", j)
			for k := 0; k < 10; k++ {
				podName := fmt.Sprintf("pod-%d-%d", j, k)
				pods = append(pods, api.Pod{
					ObjectMeta: api.ObjectMeta{
						Name: podName,
						Namespace: ns,
						Labels: map[string]string{"rcName": rcName},
					},
				})
			}
		}
	}

	for i := 0; i < b.N; i++ {
		for _, pod := range pods {
			_, err := manager.rcStore.GetPodControllers(&pod)
			if err != nil {
				fmt.Println("error: %v", err)
			}
		}
	}
}
```