@mqliang 2015-12-05
My benchmark may be a little artificial(100 namespaces, there are 10 rcs with 10 replicas in every namespace), it indicates that adding a namespace index could improve efficiency about 2 times.