@mqliang 2015-12-08
100 namespaces, taking index building time into account:
```
➜  replication git:(indexInformer) ✗ go test -v -test.bench=. -run "^BenchmarkGetPodControllers&" -test.benchmem
PASS
BenchmarkGetPodControllers-4	       3	 488857014 ns/op	96078773 B/op	 1738119 allocs/op
ok  	k8s.io/kubernetes/pkg/controller/replication	3.276s
```
100 namespaces,without taking index building time into account:

```
➜  replication git:(indexInformer) ✗ go test -v -test.bench=. -run "^BenchmarkGetPodControllers&" -test.benchmem
PASS
BenchmarkGetPodControllers-4	       3	 491981409 ns/op	95232581 B/op	 1720238 allocs/op
ok  	k8s.io/kubernetes/pkg/controller/replication	3.283s
```

Seems no big difference.

----

Results for 1000 namespaces:

Before
```
➜  replication git:(master) ✗ go test -v -test.bench=. -run "^BenchmarkGetPodControllers&" -test.benchmem
PASS
BenchmarkGetPodControllers-4	       1	126092425095 ns/op	17144683968 B/op	14209434 allocs/op
ok  	k8s.io/kubernetes/pkg/controller/replication	127.307s
```

After:
```
➜  replication git:(indexInformer) ✗ go test -v -test.bench=. -run "^BenchmarkGetPodControllers&" -test.benchmem
PASS
BenchmarkGetPodControllers-4	       1	5883614440 ns/op	952405728 B/op	17202466 allocs/op
ok  	k8s.io/kubernetes/pkg/controller/replication	7.790s
```