@mqliang 2015-12-08
@lavalamp e2e test failed. I can not access the test result. Could you kindly paste the test information?