@lavalamp 2015-12-08
Our e2es are failing a lot for unrelated reasons today. Let's run it again
later today when they're back and running.

On Mon, Dec 7, 2015 at 8:30 PM, Liang Mingqiang <notifications@github.com>
wrote:

> @lavalamp <https://github.com/lavalamp> e2e test failed. I can not access
> the test result. Could you kindly paste the test information?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/17277#issuecomment-162759776>
> .
>