@lavalamp 2015-12-09
OK, it seems it's worth doing. I have to reassign, as I'll be OOO for a bit.