@mqliang 2015-12-11
@k8s-bot test this please