@mqliang 2015-12-14
@davidopp @thockin This PR has hang about 1 month, and @bprashanth is unavailable about 1 week. Could anyone kindly help review this PR so that we can move on.

BTW, benchmark shows that adding a namespace index for cache of ReplicationControllerMamager is very helpful. I am not sure the common number of Jobs, DaemonSets(and the namespace numbers, average container numbers). In the next step, we can add namespace index for them if benchmark indicates it's worth doing in the common use cases.