@mqliang 2015-12-18
related issue: https://github.com/kubernetes/kubernetes/issues/17130