@mqliang 2015-12-18
@k8s-bot unit test this please