@bprashanth 2016-01-26
Err, sorry didn't realize this was assigned to me on account of vacation. If it's something we want for 1.2 we should probably re-assign, if not, I can prioritize after 1.2 code freeze.