@derekwaynecarr 2016-02-23
Is it possible to have this make kube 1.2?

This would be a big improvement for us in OpenShift where we tend to optimize around lots of namespaces each with a handful of RCs.

/cc @ncdc