@lavalamp 2016-03-18
OK, just getting back to my PR list. Does this really buy us anything beyond just memoizing, like we did already?