@mqliang 2016-03-19
> Does this really buy us anything beyond just memoizing, like we did already?

Usually, a rc will have a replica number 3-5, so memoizing( #19043) could only speed up `getPodController()` about 3-5 times. Adding namespace index will optimize around lots of namespaces each with a handful of RCs.