@wojtek-t 2016-03-19
@lavalamp @mqliang - so just to be clear - it really depends on the usage patterns; if you have a lot of namespaces - then this should help a lot. But the real question is how many namespaces users usually have in the cluster.
Also, regarding #19043 is depends on the numbers of pods in replication controllers - they question is whether it is really 3-5, or maybe 100?