@wojtek-t 2016-03-24
Anyway - I think that high-level it makes sense to do it.
@lavalamp - WDYT? 
If you agree that it makes sense, feel free to reassign it to me.