@mqliang 2016-03-24
Will rebase and make the same optimization to RS/Job/DS if we agree to do this. After all, we will eventually support a larger cluster, this optimizaion is very helpful.