@lavalamp 2016-03-24
OK, I'm on board with the concept. I'll let @wojtek-t review. One thing to think about, can you just add this to the existing lister? does anyone *not* want to filter by namespace? I'm not a big fan of the code duplication here :/