@mqliang 2016-03-25
@lavalamp 
>
One thing to think about, can you just add this to the existing lister?

Add this into lister.go, instead of in a new file sounds good. But `StoreToXXXLister` should still be remained since there are some place where Store do not need filter by namespace, such as NodeController and Scheduler.