@wojtek-t 2016-03-25
@mqliang - I think that what we need to do is to leave all the mthods of StoreToXXXLister as they are, but add additional method for Listing just from a namespace. Or sth like that.