@mqliang 2016-04-03
close this in flavour of https://github.com/kubernetes/kubernetes/pull/23795